Fix compiler warnings from GCC (enum bitwise, unused vars, volatile ++)
Cast InterruptID/InterruptType operands explicitly in intc.cpp
(bitwise op between different enum types is deprecated). Scope the
x86-only SSE rounding-mode temp in sh4_registers.cpp to its #if block.
Silence set-but-unused diagnostic counters in aica.cpp/arm7.cpp
(printf output stays disabled). Replace volatile-qualified ++ with
explicit assignment in wii_audio.cpp (deprecated in C++20).

+++ dc/sh4/intc.cpp
@@ -307,23 +307,23 @@ int UpdateINTC()
 
 void SetInterruptPend(InterruptID intr)
 {
-	u32 piid= intr & InterruptPIIDMask;
+	u32 piid= (u32)intr & (u32)InterruptPIIDMask;
 	interrupt_vpend|=InterruptBit[piid];
 }
 void ResetInterruptPend(InterruptID intr)
 {
-	u32 piid= intr & InterruptPIIDMask;
+	u32 piid= (u32)intr & (u32)InterruptPIIDMask;
 	interrupt_vpend&=~InterruptBit[piid];
 }
 
 void SetInterruptMask(InterruptID intr)
 {
-	u32 piid= intr & InterruptPIIDMask;
+	u32 piid= (u32)intr & (u32)InterruptPIIDMask;
 	interrupt_vmask|=InterruptBit[piid];
 }
 void ResetInterruptMask(InterruptID intr)
 {
-	u32 piid= intr & InterruptPIIDMask;
+	u32 piid= (u32)intr & (u32)InterruptPIIDMask;
 	interrupt_vmask&=~InterruptBit[piid];
 }
